backend/requirements.txt
aiosqlite
alembic
argon2-cffi
bcrypt
cachetools
cryptography
fastapi[standard]
passlib
pydantic
pydantic-settings
python-jose[cryptography]
SQLAlchemy
uvicorn